Make ui and argument to MainWindow and add wcs argument
These changes were made so that a script can pass in the ui and/or the
wcs. The ui was implemented as a global variable, which would not work
with testing without some trickery.

Providing the WCS skips astrometry.net fitting so that we avoid the need
for providing a key and so that our tests do not have to go out to
astrometry.net to be able to run.